Correct MSVC build error

+++ src/sha256.h
@@ -17,13 +17,13 @@
 
 /**************************** DATA TYPES ****************************/
 typedef unsigned char BYTE;             // 8-bit byte
-typedef unsigned int  WORD;             // 32-bit word, change to "long" for 16-bit machines
+typedef unsigned int  UINT;             // 32-bit word, change to "long" for 16-bit machines
 
 typedef struct {
 	BYTE data[64];
-	WORD datalen;
+	UINT datalen;
 	unsigned long long bitlen;
-	WORD state[8];
+	UINT state[8];
 } SHA256_CTX;
 
 /*********************** FUNCTION DECLARATIONS **********************/
